requirements.txt
pandas
playwright
customtkinter
Pillow